feat(cli): migrate run-assembly layer to internal/cli/run (#362)
Move the shared run setup — Env/SetupEnv/NewConfig plus the tool,
skills, plugin, and config-path helpers — out of cmd/pigo into
internal/cli/run so the REPL and headless drivers assemble a run
through one exported API. dispatch and the print/stream drivers stay
in cmd/pigo pending #363/#369. Tests for the moved functions migrate
to package run.
